refactor: reuse media-core chunk reader and shared dreaming session key helper
