+++ packages/memory-host-sdk/src/host/openclaw-runtime-session.ts
@@ -42,3 +42,21 @@ export function extractAgentIdFromSessionsDir(sessionsDir: string): string | nul
   }
   return parts[sessionsIndex - 1] || null;
 }
+
+/** Session-key prefix marking transcripts generated by memory dreaming runs. */
+export const DREAMING_NARRATIVE_RUN_PREFIX = "dreaming-narrative-";
+
+/** True when a session-store key belongs to a dreaming narrative run. */
+export function isDreamingNarrativeSessionStoreKey(sessionKey: string): boolean {
+  const trimmed = sessionKey.trim();
+  if (!trimmed) {
+    return false;
+  }
+  const firstSeparator = trimmed.indexOf(":");
+  if (firstSeparator < 0) {
+    return trimmed.startsWith(DREAMING_NARRATIVE_RUN_PREFIX);
+  }
+  const secondSeparator = trimmed.indexOf(":", firstSeparator + 1);
+  const sessionSegment = secondSeparator < 0 ? trimmed : trimmed.slice(secondSeparator + 1);
+  return sessionSegment.startsWith(DREAMING_NARRATIVE_RUN_PREFIX);
+}
